Refactor: Move user entity into user dir
