+++ src/party/party.controller.ts
@@ -13,7 +13,7 @@ import {
 import { Request } from 'express';
 import { Resp } from 'src/common/response';
 import { EatParty } from 'src/party/eat_party.entity';
-import { User } from 'src/entity/user.entity';
+import { User } from 'src/user/user.entity';
 import { DeleteResult } from 'typeorm';
 import { CreatePartyDto } from './party.dto';
 import { PartyService } from './party.service';
