+++ src/party/party.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { EatParty } from 'src/party/eat_party.entity';
-import { User } from 'src/entity/user.entity';
+import { User } from 'src/user/user.entity';
 import { PartyController } from './party.controller';
 import { PartyService } from './party.service';
